Fixed readme for project-management.js
