+++ .eslintrc.json
@@ -15,6 +15,8 @@
         "sourceType": "module"
     },
     "rules": {
-        "indent": ["error", 4]
+        "indent": ["error", 4],
+        "no-empty": ["error", {"allowEmptyCatch":  true}],
+        "max-len": ["error", {"ignoreComments":  true, "code":  110}]
     }
 }
